Adjust some small code
